chore: Move `StorageValues` to `toml.rs`
