+++ objects/src/accounts/component/template/storage/mod.rs
@@ -311,57 +311,6 @@ impl Deserializable for StorageEntry {
     }
 }
 
-// STORAGE VALUES
-// ================================================================================================
-
-/// Represents the type of values that can be found in a storage slot's `values` field.
-#[cfg(feature = "std")]
-#[derive(serde::Deserialize, serde::Serialize)]
-#[serde(untagged)]
-enum StorageValues {
-    /// List of individual words (for multi-slot entries).
-    Words(Vec<WordRepresentation>),
-    /// List of key-value entries (for map storage slots).
-    MapEntries(Vec<MapEntry>),
-    /// A placeholder value, represented as "{{key}}".
-    Template(StoragePlaceholder),
-}
-
-#[cfg(feature = "std")]
-impl StorageValues {
-    pub fn is_list_of_words(&self) -> bool {
-        match self {
-            StorageValues::Words(_) => true,
-            StorageValues::MapEntries(_) => false,
-            StorageValues::Template(_) => false,
-        }
-    }
-
-    pub fn into_words(self) -> Option<Vec<WordRepresentation>> {
-        match self {
-            StorageValues::Words(vec) => Some(vec),
-            StorageValues::MapEntries(_) => None,
-            StorageValues::Template(_) => None,
-        }
-    }
-
-    pub fn into_map_entries(self) -> Option<Vec<MapEntry>> {
-        match self {
-            StorageValues::Words(_) => None,
-            StorageValues::MapEntries(vec) => Some(vec),
-            StorageValues::Template(_) => None,
-        }
-    }
-
-    pub fn len(&self) -> Option<usize> {
-        match self {
-            StorageValues::Words(vec) => Some(vec.len()),
-            StorageValues::MapEntries(vec) => Some(vec.len()),
-            StorageValues::Template(_) => None,
-        }
-    }
-}
-
 // MAP ENTRY
 // ================================================================================================
